Support libpqxx version 7.x.
libpqxx 7 requires C++17 and thus requires us to tell the compiler to use
the right compiler flag. Teach autoconf to test for the proper switches
depending on the libpqxx version found.

This also defines a new macro HAVE_LIBPQXX7 in config.h, which will be
used by the follow up commit to reflect recent API changes in libpqxx 7.