Change handling of libpqxx classes to support version 7.
pqxx::connection::activate() and pqxx::connection::disconnect() are gone,
the former doesn't have a corresponding substitute. Instead the caller is
encouraged to watch out for broken_connection exception. Also the explicit
activation of a PostgreSQL connection is no longer supported.

work::prepared() is also no longer supported, so change this to
work::exec_prepared(). There's no need to special case this for older libpqxx
versions, since this method is already present there, too.

This commit wraps those deprecated API calls into a ifdef to support
builds with an older libpqxx version. We might get rid of this if we don't want
to have support for those older ones.